# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "牧民健康应用API",
    "description": "牧民健康应用的后端API接口",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API服务器"
    }
  ],
  "components": {
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "integer",
            "description": "时间戳（毫秒）"
          }
        }
      },
...
      "UserSetting": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "description": "语言设置（zh-CN/mn-MN）"
          },
          "pushNotification": {
            "type": "boolean",
            "description": "是否开启推送通知"
          },
          "soundNotification": {
            "type": "boolean",
            "description": "是否开启声音通知"
          },
          "theme": {
            "type": "string",
            "description": "主题设置（light/dark/system）"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/settings": {
      "get": {
        "tags": ["系统设置"],
        "summary": "获取系统设置",
        "description": "获取用户的系统设置",
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSetting"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "认证失败"
          }
        }
      },
      "put": {
        "tags": ["系统设置"],
        "summary": "更新系统设置",
        "description": "更新用户的系统设置",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "language": {
                    "type": "string",
                    "description": "语言设置（zh-CN/mn-MN）"
                  },
                  "pushNotification": {
                    "type": "boolean",
                    "description": "是否开启推送通知"
                  },
                  "soundNotification": {
                    "type": "boolean",
                    "description": "是否开启声音通知"
                  },
                  "theme": {
                    "type": "string",
                    "description": "主题设置（light/dark/system）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserSetting"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "认证失败"
          }
        }
      }
    },
...
  }
} 